myshop/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-bootstrap4==23.4
django-haystack==3.2.1
elastic-transport==8.12.0
elasticsearch==8.11.1
idna==3.6
pillow==10.2.0
requests==2.31.0
setuptools==69.0.3
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
